+++ Docs/CHANGELOG.md
@@ -6,3 +6,8 @@
 
 - Added `disabled` attribute to `<Catch2Adapter>` node for use in the _.runsettings_ file.
 - Added `<DebugBreak>` setting that turns on or off Catch2's break on test failure feature (`--break`), when running tests via `Debug Selected Tests`.
+
+### Extended Features
+
+ - Added `Debug` level option to `<Logging>` setting and updated what is logged in the `Verbose` level.
+